requirements.txt
boto3==1.9.4
botocore==1.12.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
cycler==0.10.0
docutils==0.14
Flask==1.0.2
future==0.16.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
line-bot-sdk==1.8.0
MarkupSafe==1.0
matplotlib==2.2.3
numpy==1.15.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
